docker/preprocessing/requirements.txt
fastapi
uvicorn
requests
Pillow
numpy
tqdm
pandas
kaggle
matplotlib
schedule
openpyxl